issue-1 ― Minor test refactor
