+++ argo-plugin/src/test/kotlin/io/github/leofuso/argo/plugin/fixtures/ResourceResolver.kt
@@ -3,6 +3,7 @@ package io.github.leofuso.argo.plugin.fixtures
 import com.speedment.common.combinatorics.Permutation
 import org.gradle.api.Project
 import org.gradle.api.file.FileTree
+import org.junit.jupiter.api.DynamicTest
 import java.io.File
 import java.lang.StackWalker.Option.RETAIN_CLASS_REFERENCE
 import java.util.*
@@ -10,7 +11,6 @@ import java.util.stream.Collectors
 import kotlin.streams.toList
 
 fun loadResource(path: String): File {
-
     val walker = StackWalker.getInstance(RETAIN_CLASS_REFERENCE)
     val callerClass = walker.callerClass
     val classLoader = callerClass.classLoader
@@ -45,3 +45,10 @@ fun permutations(project: Project, vararg files: File): Map<String, FileTree> {
 fun permutations(project: Project, tree: FileTree): Map<String, FileTree> {
     return permutations(project, *tree.files.toTypedArray())
 }
+
+fun scenarios(project: Project, tree: FileTree, executable: (FileTree) -> Unit) = permutations(project, tree)
+    .map {
+        DynamicTest.dynamicTest("Scenario ${it.key}") {
+            executable.invoke(it.value)
+        }
+    }.toList()
